Added and refactor files service
