+++ src/services/auth/get.service.ts
@@ -1,13 +1,15 @@
 //External
 const axios = require("axios");
-//Config for .dotenv
 import "dotenv/config";
+//Helpers
+import { sendPostRequest } from "../../helpers/axios/request/post";
 //Const-vars
-let axiosResponse: any;
-let tokenData: any;
 const API_PAYPAL_CREATE_TOKEN_URL: string =
-  `${process.env.API_PAYPAL_BASE_URL}:${process.env
-    .API_PAYPAL_CREATE_TOKEN_RESOURCE}` || "";
+  `${process.env.API_PAYPAL_BASE_URL}:${process.env.API_PAYPAL_CREATE_TOKEN_RESOURCE}` ||
+  "";
+let axiosResponse: any;
+let axiosData: any;
+let axiosConfig: any;
 let msgResponse: string;
 let msgLog: string;
 
@@ -19,21 +21,24 @@ let msgLog: string;
  */
 export const getAccessTokenFromPaypal = async (credentials: any) => {
   try {
-    axiosResponse = await axios.post(
-      API_PAYPAL_CREATE_TOKEN_URL,
-      new URLSearchParams({
-        grant_type: "client_credentials"
-      }),
-      {
-        auth: {
-          username: credentials.user,
-          password: credentials.password
-        }
-      }
-    );
-    tokenData = axiosResponse.data;
+    axiosData = new URLSearchParams({
+      grant_type: "client_credentials",
+    });
 
-    return tokenData;
+    axiosConfig = {
+      auth: {
+        username: credentials.user,
+        password: credentials.password,
+      },
+    };
+
+    axiosResponse = await sendPostRequest(
+      API_PAYPAL_CREATE_TOKEN_URL,
+      axiosData,
+      axiosConfig
+    );
+
+    return axiosResponse;
   } catch (error) {
     msgResponse = "ERROR in getAccessTokenFromPaypal() function controller.";
     msgLog = msgResponse + `Caused by ${error}`;
